Update unit-test CSVs. Fix ZBUffer_t I/O bug.
Add:
	test/*.csv
		-Update existing unit-tests to reflect the addition of lighting,
		amongst other features.

Fix:
	src/graphics/screen.c
		-Fix a bug in `writeZBufferToFile()` and
		`readZBufferFromFile()`, which caused `equalZBuffers()` to
		occasionally fail (incorrectly) due to floating-point errors.

+++ src/graphics/screen.c
@@ -117,7 +117,9 @@ void writeZBufferToFile(ZBuffer_t *zBuf, const char *filePath){
 	int y, x;
 	for(y = 0; y < IMAGE_HEIGHT; y++)
 		for(x = 0; x < IMAGE_WIDTH; x++)
-			fprintf(file, "%lf,%lf,", zBuf->buf[y][x][0], zBuf->buf[y][x][1]);
+			fprintf(
+					file, "%d,%d,", (int)zBuf->buf[y][x][0],
+					(int)zBuf->buf[y][x][1]);
 
 	fclose(file);
 }
